🏗️ move resume page into its own component
